// src/common/Enums.ts
export enum RolerUser {
  ADMIN = 'ADMIN',
  MOD = 'MOD',
  MEM = 'MEM',
}

export enum LEN_OF_FIELDS {
  LENGTH_LOW = 100,
  LENGTH_MEDIUM = 250,
  LENGTH_HEIGHT = 500,
}

export enum IMAGE_UPLOAD_TYPE {
  SERVER = 'SERVER',
  IPFS = 'IPFS',
  IMGUR_COM = 'IMGUR_COM',
  FACEBOOK_AVATAR = 'FACEBOOK_AVATAR',
  GOOGLE_AVATAR = 'GOOGLE_AVATAR',
}

export enum QUEUE_NAME {
  SEND_MAIL = 'send-mail',
}

export enum COUNT_TYPE {
  SEND_MAIL = 'SEND_MAIL',
}

export enum REACT_TYPE {
  EXAM_COMMIT = 'EXAM_COMMIT',
  USER_LIKE = 'USER_LIKE',
}

export enum COMMENT_REACT_TYPE {
  NOTHING = 'NOTHING',
  LIKE = 'LIKE',
  DISLIKE = 'DISLIKE',
}

export enum SEX_USER {
  MALE = 'MALE',
  FEMALE = 'FEMALE',
  OTHER = 'OTHER',
}